Remove the top-level tests module completely.
